# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loop Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationDetailType } from './conversationDetailType';
import type { ConversationMember } from './conversationMember';

export interface ConversationDetail {
  id: number;
  type: ConversationDetailType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatar?: string | null;
  members: ConversationMember[];
  createdAt: string;
}