# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loop Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationInputType } from './conversationInputType';

export interface ConversationInput {
  type: ConversationInputType;
  /** @nullable */
  name?: string | null;
  memberIds: number[];
}